refactor(element3): move Divider to src/components
