+++ packages/element3/packages/theme-chalk/src/index.scss
@@ -22,7 +22,6 @@
 @import "./container.scss";
 @import "./date-picker.scss";
 @import "./dialog.scss";
-@import "./divider.scss";
 @import "./drawer.scss";
 @import "./dropdown.scss";
 @import "./dropdown-item.scss";
@@ -75,6 +74,7 @@
 @import "./badge.scss";
 @import "./button.scss";
 @import "./container.scss";
+@import "./divider.scss";
 @import "./link.scss";
 @import "./message.scss";
 @import "./switch.scss";
