+++ packages/element3/src/components/Divider/index.js
@@ -1,8 +1,8 @@
-import ElDivider from './Divider'
+import ElDivider from './src/Divider'
 
 /* istanbul ignore next */
 ElDivider.install = function (app) {
   app.component(ElDivider.name, ElDivider)
 }
 
-export default ElDivider
+export { ElDivider }
