+++ packages/element3/src/index.js
@@ -83,7 +83,7 @@ import ElCollapse from '../packages/collapse'
 import ElCollapseItem from '../packages/collapse-item'
 import ElTimeline from '../packages/timeline'
 import ElTimelineItem from '../packages/timeline-item'
-import ElDivider from '../packages/divider'
+import { ElDivider } from './components/Divider'
 import ElCalendar from '../packages/calendar'
 import ElImage from '../packages/image'
 import ElBacktop from '../packages/backtop'
